# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TAU Service API",
    "description": "TAU Service Template API documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user",
        "description": "Returns the currently authenticated user's profile",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/contracts": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "List contracts",
        "description": "Returns a paginated, filtered list of contracts with deadline aggregate counts. Open to any authenticated user.",
        "operationId": "listContracts",
        "parameters": [
          {
            "name": "fy",
            "in": "query",
            "description": "Filter by fiscal year (e.g. FY26)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Free-text search across contract fields",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size",
            "schema": {
              "type": "integer",
              "default": 30,
              "minimum": 1
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Sort field: pop_end or execution_quarter",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_dir",
            "in": "query",
            "description": "Sort direction: asc or desc",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contracts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListContractsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Create a contract",
        "description": "Creates a new contract. Requires admin or editor role.",
        "operationId": "createContract",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContractRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contract created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateContractResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin or editor role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/fiscal-years": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "List distinct fiscal years",
        "description": "Returns the distinct fiscal years present across contracts, for use in the filter dropdown. Open to any authenticated user.",
        "operationId": "listFiscalYears",
        "responses": {
          "200": {
            "description": "Fiscal years retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListFiscalYearsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/{id}": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Get contract by ID",
        "description": "Returns a single contract by its ID. Open to any authenticated user.",
        "operationId": "getContract",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Contract ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contract retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetContractResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Contracts"
        ],
        "summary": "Update a contract",
        "description": "Partially updates a contract. Only provided fields are changed. Requires admin or editor role.",
        "operationId": "updateContract",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Contract ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContractRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contract updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateContractResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin or editor role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contracts"
        ],
        "summary": "Delete a contract",
        "description": "Deletes a contract. Requires admin or editor role.",
        "operationId": "deleteContract",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Contract ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Contract deleted successfully"
          },
          "400": {
            "description": "Invalid ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin or editor role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/export/contracts": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Export contracts as CSV",
        "description": "Streams a filtered CSV of contracts. Read-only; open to any authenticated user.",
        "operationId": "exportContracts",
        "parameters": [
          {
            "name": "fy",
            "in": "query",
            "description": "Comma-separated fiscal years to include (empty = all)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "columns",
            "in": "query",
            "description": "Comma-separated column subset (empty = all, canonical order)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file returned",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "User's unique identifier"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "name": {
            "type": "string",
            "description": "User's display name"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User's assigned roles"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesResponse"
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last login timestamp"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "roles",
          "preferences",
          "created_at",
          "updated_at"
        ]
      },
      "PreferencesResponse": {
        "type": "object",
        "properties": {
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark"
            ],
            "description": "User's preferred theme"
          }
        },
        "required": [
          "theme"
        ]
      },
      "GetUserResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/UserResponse"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code for programmatic handling"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "description": "Validation error details (only present for validation errors)"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Field that failed validation"
          },
          "message": {
            "type": "string",
            "description": "Validation error message"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
...
      "ContractResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Contract unique identifier",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Contract title"
          },
          "company": {
            "type": "string",
            "description": "Company / vendor name"
          },
          "poc_name": {
            "type": "string",
            "nullable": true,
            "description": "Point-of-contact name"
          },
          "poc_email": {
            "type": "string",
            "nullable": true,
            "description": "Point-of-contact email"
          },
          "poc_phone": {
            "type": "string",
            "nullable": true,
            "description": "Point-of-contact phone"
          },
          "pop_start": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Period-of-performance start (YYYY-MM-DD)"
          },
          "pop_end": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Period-of-performance end (YYYY-MM-DD)"
          },
          "execution_quarter": {
            "type": "string",
            "nullable": true,
            "enum": [
              "Q1",
              "Q2",
              "Q3",
              "Q4"
            ],
            "description": "Execution quarter"
          },
          "fiscal_year": {
            "type": "string",
            "description": "Fiscal year (e.g. FY26)"
          },
          "notes": {
            "type": "string",
            "description": "Free-form notes"
          },
          "logform_number": {
            "type": "string",
            "nullable": true,
            "description": "Logform number"
          },
          "logform_url": {
            "type": "string",
            "nullable": true,
            "description": "Logform URL"
          },
          "updated_by": {
            "type": "string",
            "description": "Display name of last editor"
          },
          "created_at": {
            "type": "string",
            "description": "Creation timestamp",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "description": "Last update timestamp",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "company",
          "fiscal_year",
          "notes",
          "updated_by",
          "created_at",
          "updated_at"
        ]
      },
      "ContractCountsResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total contracts matching the filter"
          },
          "expiring_30": {
            "type": "integer",
            "description": "Count with pop_end within 30 days"
          },
          "expiring_60": {
            "type": "integer",
            "description": "Count with pop_end in 31-60 days"
          },
          "expiring_90": {
            "type": "integer",
            "description": "Count with pop_end in 61-90 days"
          }
        },
        "required": [
          "total",
          "expiring_30",
          "expiring_60",
          "expiring_90"
        ]
      },
      "CreateContractRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Contract title",
            "minLength": 1,
            "maxLength": 200
          },
          "company": {
            "type": "string",
            "description": "Company / vendor name",
            "minLength": 1,
            "maxLength": 200
          },
          "poc_name": {
            "type": "string",
            "nullable": true,
            "description": "Point-of-contact name"
          },
          "poc_email": {
            "type": "string",
            "nullable": true,
            "description": "Point-of-contact email"
          },
          "poc_phone": {
            "type": "string",
            "nullable": true,
            "description": "Point-of-contact phone"
          },
          "pop_start": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Period-of-performance start (YYYY-MM-DD)"
          },
          "pop_end": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Period-of-performance end (YYYY-MM-DD)"
          },
          "execution_quarter": {
            "type": "string",
            "nullable": true,
            "enum": [
              "Q1",
              "Q2",
              "Q3",
              "Q4"
            ],
            "description": "Execution quarter"
          },
          "fiscal_year": {
            "type": "string",
            "description": "Fiscal year (e.g. FY26)",
            "minLength": 1,
            "maxLength": 4
          },
          "notes": {
            "type": "string",
            "description": "Free-form notes",
            "maxLength": 250
          },
          "logform_number": {
            "type": "string",
            "description": "Logform number",
            "maxLength": 20
          },
          "logform_url": {
            "type": "string",
            "description": "Logform URL",
            "maxLength": 2000
          }
        },
        "required": [
          "title",
          "company",
          "fiscal_year"
        ]
      },
      "UpdateContractRequest": {
        "type": "object",
        "description": "Partial update. Only provided fields are changed.",
        "properties": {
          "title": {
            "type": "string",
            "description": "Contract title",
            "minLength": 1,
            "maxLength": 200
          },
          "company": {
            "type": "string",
            "description": "Company / vendor name",
            "minLength": 1,
            "maxLength": 200
          },
          "poc_name": {
            "type": "string",
            "nullable": true,
            "description": "Point-of-contact name"
          },
          "poc_email": {
            "type": "string",
            "nullable": true,
            "description": "Point-of-contact email"
          },
          "poc_phone": {
            "type": "string",
            "nullable": true,
            "description": "Point-of-contact phone"
          },
          "pop_start": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Period-of-performance start (YYYY-MM-DD)"
          },
          "pop_end": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Period-of-performance end (YYYY-MM-DD)"
          },
          "execution_quarter": {
            "type": "string",
            "nullable": true,
            "enum": [
              "Q1",
              "Q2",
              "Q3",
              "Q4"
            ],
            "description": "Execution quarter"
          },
          "fiscal_year": {
            "type": "string",
            "description": "Fiscal year (e.g. FY26)",
            "minLength": 1,
            "maxLength": 4
          },
          "notes": {
            "type": "string",
            "description": "Free-form notes",
            "maxLength": 250
          },
          "logform_number": {
            "type": "string",
            "description": "Logform number",
            "maxLength": 20
          },
          "logform_url": {
            "type": "string",
            "description": "Logform URL",
            "maxLength": 2000
          }
        }
      },
      "CreateContractResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "contract": {
                "$ref": "#/components/schemas/ContractResponse"
              }
            }
          }
        }
      },
      "GetContractResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "contract": {
                "$ref": "#/components/schemas/ContractResponse"
              }
            }
          }
        }
      },
      "UpdateContractResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "contract": {
                "$ref": "#/components/schemas/ContractResponse"
              }
            }
          }
        }
      },
      "ListContractsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "contracts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ContractResponse"
                }
              },
              "total": {
                "type": "integer",
                "description": "Total contracts matching the filter"
              },
              "page": {
                "type": "integer",
                "description": "Current page (1-based)"
              },
              "total_pages": {
                "type": "integer",
                "description": "Total number of pages"
              },
              "counts": {
                "$ref": "#/components/schemas/ContractCountsResponse"
              }
            }
          }
        }
      },
      "ListFiscalYearsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "fiscal_years": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Distinct fiscal years"
              }
            }
          }
        }
      },
...
    }
  },
...
}